Rename editor sync entry to publish
