+++ assets/i18n/chs.js
@@ -124,10 +124,10 @@ const translations = {
         system: '系统',
         siteSettings: '站点设置',
         nanoSiteUpdates: 'NanoSite 更新',
-        sync: '同步',
+        sync: '发布',
         siteSettingsMeta: '编辑 site.yaml 设置。',
         systemUpdatesMeta: '查看并应用 NanoSite 更新。',
-        syncMeta: '查看本地与 GitHub 的同步状态。',
+        syncMeta: '将本地更改发布到 GitHub。',
         editorLanguage: '编辑器语言',
         editorLanguageMeta: '调整编辑器界面语言。',
         languages: '种语言',
