+++ assets/i18n/cht-tw.js
@@ -124,10 +124,10 @@ const translations = {
         system: '系統',
         siteSettings: '站點設定',
         nanoSiteUpdates: 'NanoSite 更新',
-        sync: '同步',
+        sync: '發布',
         siteSettingsMeta: '編輯 site.yaml 設定。',
         systemUpdatesMeta: '檢視並套用 NanoSite 更新。',
-        syncMeta: '檢視本地與 GitHub 的同步狀態。',
+        syncMeta: '將本地更改發布到 GitHub。',
         editorLanguage: '編輯器語言',
         editorLanguageMeta: '調整編輯器介面語言。',
         languages: '種語言',
