+++ assets/i18n/en.js
@@ -124,10 +124,10 @@ const translations = {
         system: 'System',
         siteSettings: 'Site Settings',
         nanoSiteUpdates: 'NanoSite Updates',
-        sync: 'Sync',
+        sync: 'Publish',
         siteSettingsMeta: 'Edit site.yaml settings.',
         systemUpdatesMeta: 'Review and apply NanoSite updates.',
-        syncMeta: 'Review local and GitHub sync status.',
+        syncMeta: 'Publish local changes to GitHub.',
         editorLanguage: 'Editor language',
         editorLanguageMeta: 'Change the editor interface language.',
         languages: 'languages',
