+++ assets/i18n/ja.js
@@ -124,10 +124,10 @@ const translations = {
         system: 'システム',
         siteSettings: 'サイト設定',
         nanoSiteUpdates: 'NanoSite 更新',
-        sync: '同期',
+        sync: '公開',
         siteSettingsMeta: 'site.yaml 設定を編集します。',
         systemUpdatesMeta: 'NanoSite 更新を確認して適用します。',
-        syncMeta: 'ローカルと GitHub の同期状態を確認します。',
+        syncMeta: 'ローカルの変更を GitHub に公開します。',
         editorLanguage: 'エディター言語',
         editorLanguageMeta: 'エディター UI の言語を変更します。',
         languages: '言語',
