+++ assets/i18n/languages.json
@@ -1,7 +1,7 @@
 [
-  { "value": "en", "label": "English", "module": "./en.js?v=20260504i18n" },
-  { "value": "chs", "label": "简体中文", "module": "./chs.js?v=20260504i18n" },
-  { "value": "cht-tw", "label": "正體中文（台灣）", "module": "./cht-tw.js?v=20260504i18n" },
-  { "value": "cht-hk", "label": "繁體中文（香港）", "module": "./cht-hk.js?v=20260504i18n" },
-  { "value": "ja", "label": "日本語", "module": "./ja.js?v=20260504i18n" }
+  { "value": "en", "label": "English", "module": "./en.js?v=20260504publish" },
+  { "value": "chs", "label": "简体中文", "module": "./chs.js?v=20260504publish" },
+  { "value": "cht-tw", "label": "正體中文（台灣）", "module": "./cht-tw.js?v=20260504publish" },
+  { "value": "cht-hk", "label": "繁體中文（香港）", "module": "./cht-hk.js?v=20260504publish" },
+  { "value": "ja", "label": "日本語", "module": "./ja.js?v=20260504publish" }
 ]
