+++ assets/js/i18n.js
@@ -12,7 +12,7 @@
 import { parseFrontMatter } from './content.js';
 import { getContentRoot } from './utils.js';
 import { fetchConfigWithYamlFallback } from './yaml.js';
-import enTranslations, { languageMeta as enLanguageMeta } from '../i18n/en.js?v=20260504i18n';
+import enTranslations, { languageMeta as enLanguageMeta } from '../i18n/en.js?v=20260504publish';
 
 // Content fetch cache modes are normalized by cache-control.js.
