+++ index_editor.html
@@ -2210,9 +2210,9 @@
       updateOffset();
     })();
   </script>
-  <script type="module" src="assets/js/editor-boot.js?v=editor-i18n-20260504"></script>
-  <script type="module" src="assets/js/editor-main.js?v=editor-block-reorder-toolbar-sync-20260504"></script>
-  <script type="module" src="assets/js/composer.js?v=editor-i18n-20260504"></script>
+  <script type="module" src="assets/js/editor-boot.js?v=editor-publish-label-20260504"></script>
+  <script type="module" src="assets/js/editor-main.js?v=editor-publish-label-20260504"></script>
+  <script type="module" src="assets/js/composer.js?v=editor-publish-label-20260504"></script>
   <!-- Back to top button -->
   <button id="backToTop" class="back-to-top" aria-label="返回顶部" title="返回顶部" type="button" hidden>
     <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" aria-hidden="true">
